+++ NEWS.md
@@ -1,6 +1,7 @@
 # suppoRt 0.5
   * New function `calc_cv()`
   * New function `as_data_table()`
+  * New function `setup_project()`
 
 # suppoRt 0.4.2
   * use GitHub actions